Added utility for pulling random sentences from an article
As per Rachel's request, rand_sents.py will pull N random sentences from an
article file and write them to an output file.  Defaults: N = 10, pull from
data/set1/a2.htm, and write to src/test_files/rand_sents.txt.  The comments
at the beginning of the script explain how to easily change the defaults,
either in the file or on the command line.  (Run it in the src/ directory!)